make vision handling dedicated class + set food property method back to private + remove unneeded parameter to food property building (default is 0)
