+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/VisionHandler.java
@@ -1,26 +1,25 @@
 package by.dragonsurvivalteam.dragonsurvival.client.render;
 
-import by.dragonsurvivalteam.dragonsurvival.client.util.ClientUtils;
-import by.dragonsurvivalteam.dragonsurvival.common.capability.DragonStateProvider;
-import by.dragonsurvivalteam.dragonsurvival.common.dragon_types.DragonTypes;
-import by.dragonsurvivalteam.dragonsurvival.util.DragonUtils;
+import by.dragonsurvivalteam.dragonsurvival.registry.DSEffects;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.world.level.material.FogType;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.neoforge.client.event.RenderBlockScreenEffectEvent;
+import net.neoforged.fml.loading.FMLEnvironment;
 import net.neoforged.neoforge.client.event.RenderLevelStageEvent;
 import net.neoforged.neoforge.client.event.ViewportEvent;
 
+/** Handles water and lava vision effects */
 @EventBusSubscriber(Dist.CLIENT)
-public class ClientEvents {
+public class VisionHandler {
     private static boolean hasUpdatedSinceChangingLavaVision = false;
     private static boolean hasLavaVisionPrev = false;
 
     @SubscribeEvent
     public static void onRenderFog(ViewportEvent.RenderFog event) {
-        if (ClientUtils.hasLavaVision() && event.getCamera().getFluidInCamera() == FogType.LAVA) {
+        if (hasLavaVision() && event.getCamera().getFluidInCamera() == FogType.LAVA) {
             event.setNearPlaneDistance(0);
             event.setFarPlaneDistance(event.getRenderer().getRenderDistance() * 0.2f);
             event.setCanceled(true);
@@ -33,7 +32,7 @@ public class ClientEvents {
             return;
         }
 
-        if (ClientUtils.hasLavaVision()) {
+        if (hasLavaVision()) {
             if (!hasLavaVisionPrev) {
                 hasUpdatedSinceChangingLavaVision = false;
             }
@@ -57,9 +56,9 @@ public class ClientEvents {
     public static void onRenderFog(ViewportEvent.ComputeFogColor event) {
         float adjustment = 1;
 
-        if (ClientUtils.hasLavaVision() && event.getCamera().getFluidInCamera() == FogType.LAVA) {
+        if (hasLavaVision() && event.getCamera().getFluidInCamera() == FogType.LAVA) {
             adjustment = 0.15f;
-        } else if (ClientUtils.hasWaterVision() && event.getCamera().getFluidInCamera() == FogType.WATER) {
+        } else if (hasWaterVision() && event.getCamera().getFluidInCamera() == FogType.WATER) {
             adjustment = 0.25f;
         }
 
@@ -68,16 +67,31 @@ public class ClientEvents {
         event.setGreen(event.getGreen() * adjustment);
     }
 
-    @SubscribeEvent
-    public static void removeFireOverlay(RenderBlockScreenEffectEvent event) {
-        if (event.getOverlayType() != RenderBlockScreenEffectEvent.OverlayType.FIRE) {
-            return;
+    public static boolean hasLavaVision() {
+        if (FMLEnvironment.dist != Dist.CLIENT) {
+            return false;
         }
 
-        DragonStateProvider.getCap(Minecraft.getInstance().player).ifPresent(handler -> {
-            if (DragonUtils.isDragonType(handler, DragonTypes.CAVE)) {
-                event.setCanceled(true);
-            }
-        });
+        LocalPlayer player = Minecraft.getInstance().player;
+
+        if (player != null) {
+            return player.hasEffect(DSEffects.LAVA_VISION);
+        }
+
+        return false;
+    }
+
+    public static boolean hasWaterVision() {
+        if (FMLEnvironment.dist != Dist.CLIENT) {
+            return false;
+        }
+
+        LocalPlayer player = Minecraft.getInstance().player;
+
+        if (player != null) {
+            return player.hasEffect(DSEffects.WATER_VISION);
+        }
+
+        return false;
     }
 }
