+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/mixins/client/LightTextureMixin.java
@@ -1,6 +1,6 @@
 package by.dragonsurvivalteam.dragonsurvival.mixins.client;
 
-import by.dragonsurvivalteam.dragonsurvival.client.util.ClientUtils;
+import by.dragonsurvivalteam.dragonsurvival.client.render.VisionHandler;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import net.minecraft.client.renderer.LightTexture;
 import org.spongepowered.asm.mixin.Mixin;
@@ -11,6 +11,6 @@ import org.spongepowered.asm.mixin.injection.At;
 public abstract class LightTextureMixin {
     @ModifyExpressionValue(method = "updateLightTexture", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/player/LocalPlayer;hasEffect(Lnet/minecraft/core/Holder;)Z", ordinal = 1))
     private boolean dragonSurvival$handleWaterVision(boolean original) {
-        return original || ClientUtils.hasWaterVision();
+        return original || VisionHandler.hasWaterVision();
     }
 }
