+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/mixins/client/LiquidBlockRendererMixin.java
@@ -1,6 +1,6 @@
 package by.dragonsurvivalteam.dragonsurvival.mixins.client;
 
-import by.dragonsurvivalteam.dragonsurvival.client.util.ClientUtils;
+import by.dragonsurvivalteam.dragonsurvival.client.render.VisionHandler;
 import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.client.renderer.block.LiquidBlockRenderer;
 import net.minecraft.tags.FluidTags;
@@ -14,7 +14,7 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 public abstract class LiquidBlockRendererMixin {
     @ModifyVariable(method = "tesselate", at = @At(value = "STORE"), ordinal = 0)
     private float dragonSurvival$handleVision(float alpha, @Local(argsOnly = true) FluidState fluid, @Local(ordinal = 0) boolean isLava) {
-        if (isLava && ClientUtils.hasLavaVision() || ClientUtils.hasWaterVision() && fluid.is(FluidTags.WATER)) {
+        if (isLava && VisionHandler.hasLavaVision() || VisionHandler.hasWaterVision() && fluid.is(FluidTags.WATER)) {
             return alpha * 0.35f;
         }
