+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/mixins/client/LocalPlayerMixin.java
@@ -1,6 +1,6 @@
 package by.dragonsurvivalteam.dragonsurvival.mixins.client;
 
-import by.dragonsurvivalteam.dragonsurvival.client.util.ClientUtils;
+import by.dragonsurvivalteam.dragonsurvival.client.render.VisionHandler;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import net.minecraft.client.player.LocalPlayer;
 import org.spongepowered.asm.mixin.Mixin;
@@ -11,7 +11,7 @@ import org.spongepowered.asm.mixin.injection.At;
 public abstract class LocalPlayerMixin {
     @ModifyExpressionValue(method = "getWaterVision", at = @At(value = "FIELD", target = "Lnet/minecraft/client/player/LocalPlayer;waterVisionTime:I", ordinal = 0))
     private int dragonSurvival$handleWaterVision(int original) {
-        if (ClientUtils.hasWaterVision()) {
+        if (VisionHandler.hasWaterVision()) {
             return 600;
         }
